requirements.txt
pymupdf==1.24.5
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
sentence-transformers==3.0.1
faiss-cpu==1.8.0
streamlit==1.37.1
python-dotenv==1.0.1
torch==2.3.1
transformers==4.43.3
numpy==1.26.4